pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repofail"
version = "0.1.0"
description = "Predict why a repository will fail on your machine before you run it."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "tomli>=2.0; python_version < '3.11'",
    "pyyaml>=6.0",
]

[project.scripts]
repofail = "repofail.cli:_main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

